fix: invalid uint syntax

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,16 +7,17 @@
  * Return: number of digits
  */
 
-int main (void)
+int main(void)
 {
 	printf("1, 2, ");
-	unsigned int x=1, y=2;
+	long x=1, y=2;
 	for (int i=2;i<50;i++)
 	{
-		unsigned int z = x+y;
+		long z = x+y;
 		printf("%i, ", z);
 		x=y;	y=z;
 	}
+
 	printf('\n');
 	return 0;
 }
